relocate Bodu.Text.Encoding-only contract and KATs out of Bodu.Test
One contract base and two KAT records were consumed only by
Bodu.Text.Encoding.Test, so they move alongside their consumers per the
single-consumer rule.

Moved to Bodu.Text.Encoding/test/Text.Encoding/Contracts/ (namespace
Bodu.Text.Encoding.Contracts, matching existing concrete subclasses):

Contract:
- BinaryEncodingContractTests<TEncoding>

KATs:
- BinaryEncodingKat
- InvalidEncodedTextKat

Subclass `using` updates (6 files in Text.Encoding/Contracts/):
- Removed `using Bodu.Test.Contracts;` and `using Bodu.Test.Kat;`
  (everything now in same namespace as bases and KATs).

EncodingKnownAnswerVector (in Bodu.Text.Encoding/test/Infrastructure/)
inspected but NOT consolidated. It has richer fields than the
relocated BinaryEncodingKat: a Source citation string surfaced in
ToString(), plus FromAscii() / FromHex() factory helpers for RFC 4648
and Bitcoin Core fixture inputs. The BinaryEncodingKat XML doc
itself documents both records as coexisting alternatives ("choice is
one of convenience rather than capability"), so consolidation is not
warranted.

Verified: dotnet build passes 0 errors;
dotnet test --settings bvt.runsettings passes 1609/1609.
