+++ Bodu.Text.Encoding/test/Text.Encoding/Contracts/Base16ContractTests.cs
@@ -4,8 +4,6 @@
 // </copyright>
 // ---------------------------------------------------------------------------------------------------------------
 
-using Bodu.Test.Contracts;
-using Bodu.Test.Kat;
 
 namespace Bodu.Text.Encoding.Contracts;
