+++ Bodu.Text.Encoding/test/Text.Encoding/Contracts/InvalidEncodedTextKat.cs
@@ -1,10 +1,12 @@
-﻿// ---------------------------------------------------------------------------------------------------------------
+// ---------------------------------------------------------------------------------------------------------------
 // <copyright file="InvalidEncodedTextKat.cs" company="Bodu Pty. Ltd.">
 //     Copyright (c) Bodu Pty. Ltd.. All rights reserved.
 // </copyright>
 // ---------------------------------------------------------------------------------------------------------------
 
-namespace Bodu.Test.Kat;
+using Bodu.Test.Kat;
+
+namespace Bodu.Text.Encoding.Contracts;
 
 /// <summary>
 /// Represents a known-answer test row for malformed encoded text — the input that the decoder is expected to reject —
